+++ backend/src/services/mutualfund.rs
@@ -2,9 +2,9 @@ use crate::errors::ApiError;
 use crate::models::common::BulkCreateResponse;
 use crate::models::csv_import::{CsvPreviewResponse, CsvRowError, CsvUploadResponse};
 use crate::models::mutualfund::{CreateMutualfundRequest, Mutualfund};
-use crate::services::csv_import::{
-    build_preview, compute_taxes, finish_csv_upload, get_cell, parse_csv, parse_required_date,
-    parse_required_number, parse_required_string,
+use crate::services::csv_import::{build_preview, finish_csv_upload, parse_csv};
+use crate::services::csv_parse::{
+    compute_taxes, get_cell, parse_required_date, parse_required_number, parse_required_string,
 };
 use sqlx::PgPool;
 use std::collections::HashMap;
